move paths to installer
