+++ pkg/fleet/installer/packages/datadog_agent_windows.go
@@ -14,7 +14,7 @@ import (
 	"path"
 
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/msi"
-	"github.com/DataDog/datadog-agent/pkg/fleet/internal/paths"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/paths"
 	"github.com/DataDog/datadog-agent/pkg/fleet/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"golang.org/x/sys/windows/registry"
