+++ pkg/fleet/installer/packages/datadog_installer_windows.go
@@ -15,7 +15,7 @@ import (
 	"path"
 
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/msi"
-	"github.com/DataDog/datadog-agent/pkg/fleet/internal/paths"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/paths"
 )
 
 const (
