+++ pkg/fleet/installer/paths/installer_paths.go
