+++ pkg/fleet/installer/paths/installer_paths_windows.go
