+++ pkg/fleet/installer/paths/paths_windows_test.go
